Do not start threaded jobs automatically
If a threaded job is started automatically right after it is constructed, the
code using it has no chance to connect the "completed" signal(s) without risking
that the job function run in the thread may already be finished when the
signal(s) is/are connected which may result in the signal handler(s) never being
called. We need to prevent such race condition.